Adjust on replicator throttle  and iodepth

+++ pfs/src/pf_restful_api.cpp
@@ -504,7 +504,7 @@ void handle_save_md_disk(struct mg_connection *nc, struct http_message * hm) {
 void handle_stat_conn(struct mg_connection* nc, struct http_message* hm) {
 	std::string rst = format_string("established:%d closed:%d released:%d", PfConnection::total_count, PfConnection::closed_count, PfConnection::released_count);
 	mg_send_head(nc, 200, rst.size(), "Content-Type: text/plain");
-	mg_send(nc, rst.c_str(), rst.length());
+	mg_send(nc, rst.c_str(), (int)rst.length());
 }
 
 void handle_get_snap_list(struct mg_connection *nc, struct http_message * hm) {
